+++ utils/id_test.go
@@ -50,6 +50,74 @@ func Test_IsResourceId(t *testing.T) {
 	}
 }
 
+func Test_ActionName(t *testing.T) {
+	testcases := []struct {
+		Input  string
+		Expect string
+	}{
+		{
+			Input:  "/subscriptions/00000000-0000-0000-0000-000000000000",
+			Expect: "",
+		},
+		{
+			Input:  "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/rg",
+			Expect: "",
+		},
+		{
+			Input:  "/",
+			Expect: "",
+		},
+		{
+			Input:  "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/rg/providers/Microsoft.Automation/automationAccounts/aa/runbooks/rb/providers/Microsoft.Resources/locks/myLock",
+			Expect: "",
+		},
+		{
+			Input:  "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/rg/providers/Microsoft.Automation/automationAccounts/aa/runbooks/rb/start",
+			Expect: "start",
+		},
+	}
+
+	for _, testcase := range testcases {
+		t.Logf("[DEBUG] testcase: %s", testcase.Input)
+		actual := utils.ActionName(testcase.Input)
+		if actual != testcase.Expect {
+			t.Fatalf("[ERROR] expect %s, actual %s", testcase.Expect, actual)
+		}
+	}
+}
+
+func Test_LastSegment(t *testing.T) {
+	testcases := []struct {
+		Input  string
+		Expect string
+	}{
+		{
+			Input:  "/subscriptions/00000000-0000-0000-0000-000000000000",
+			Expect: "00000000-0000-0000-0000-000000000000",
+		},
+		{
+			Input:  "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/rg",
+			Expect: "rg",
+		},
+		{
+			Input:  "/",
+			Expect: "",
+		},
+		{
+			Input:  "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/rg/validate",
+			Expect: "validate",
+		},
+	}
+
+	for _, testcase := range testcases {
+		t.Logf("[DEBUG] testcase: %s", testcase.Input)
+		actual := utils.LastSegment(testcase.Input)
+		if actual != testcase.Expect {
+			t.Fatalf("[ERROR] expect %s, actual %s", testcase.Expect, actual)
+		}
+	}
+}
+
 func Test_ResourceTypeOfResourceId(t *testing.T) {
 	testcases := []struct {
 		Input  string
@@ -97,11 +165,51 @@ func Test_ResourceTypeOfResourceId(t *testing.T) {
 			Input:  "/subscriptions/00000000-0000-0000-0000-000000000000/providers/Microsoft.Automation/checkNameAvailability",
 			Expect: "Microsoft.Automation",
 		},
+		{
+			Input:  "/subscriptions/00000000-0000-0000-0000-000000000000/providers/Microsoft.Automation/automationAccounts/aa/runbooks/rb/providers/Microsoft.Resources/locks/myLock",
+			Expect: "Microsoft.Resources/locks",
+		},
 	}
 
 	for _, testcase := range testcases {
 		t.Logf("[DEBUG] testcase: %s", testcase.Input)
 		actual := utils.ResourceTypeOfResourceId(testcase.Input)
+		if actual != testcase.Expect {
+			t.Fatalf("[ERROR] expect %s, actual %s", testcase.Expect, actual)
+		}
+	}
+}
+
+func Test_ScopeOfListAction(t *testing.T) {
+	testcases := []struct {
+		Input  string
+		Expect string
+	}{
+		{
+			Input:  "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups",
+			Expect: "/subscriptions/00000000-0000-0000-0000-000000000000",
+		},
+		{
+			Input:  "/",
+			Expect: "",
+		},
+		{
+			Input:  "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/rg/providers/Microsoft.Automation/automationAccounts/aa/runbooks/rb/providers/Microsoft.Resources/locks",
+			Expect: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/rg/providers/Microsoft.Automation/automationAccounts/aa/runbooks/rb",
+		},
+		{
+			Input:  "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/rg/providers/Microsoft.Automation/automationAccounts/aa/runbooks",
+			Expect: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/rg/providers/Microsoft.Automation/automationAccounts/aa",
+		},
+		{
+			Input:  "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/rg/providers/Microsoft.Automation/automationAccounts",
+			Expect: "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/rg",
+		},
+	}
+
+	for _, testcase := range testcases {
+		t.Logf("[DEBUG] testcase: %s", testcase.Input)
+		actual := utils.ScopeOfListAction(testcase.Input)
 		if actual != testcase.Expect {
 			t.Fatalf("[ERROR] expect %s, actual %s", testcase.Expect, actual)
 		}